missing docs. stakepool -> vsp. other fixes
